Revert "feat: cleaning up repo and setting standards"
