--- .github/pull_request_template.md
@@ -1,18 +1,0 @@
-# Description
-
-Please include a summary of the change and list any dependencies that are required for this change
-
-# Jira Ticket
-
-Please add a link to the related Jira ticket
-Please also add a link to this PR in the Jira ticket
-
-# Checklist
-
-These are the items (excluding github checks) that should be confirmed before a branch us ready to merge
-
-- Acceptance Criteria met
-- Code has been self-reviewed and manually tested
-- Added tests that prove my fix is effective and that my feature is working
-- Test coverage of new code meets 80%
-- Added / updated any relevant documentation
