+++ ohana-homes/src/App.js
@@ -1,0 +1,29 @@
+import logo from './logo.svg';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+        <p>
+          You lookin' skinnier, and sexier, by God. You look like a whole-ass snack, straight up bussin For real?
+On God, on a stack, no cap.
+        </p>
+      </header>
+    </div>
+  );
+}
+
+export default App;
