--- src/Pages/home/components/About.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-const About = ({ props }) => {
-  return (
-    <section>
-      <div color="black">
-        <h1>About Section of Home Page</h1>
-      </div>
-    </section>
-  );
-};
-
-export default About;
