--- src/Pages/home/components/Hero.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-const Hero = ({ props }) => {
-  return (
-    <section>
-      <div color="black">
-        <h1>Hero Section of Home Page</h1>
-      </div>
-    </section>
-  );
-};
-
-export default Hero;
