--- src/Pages/islands/components/Islands.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-const Islands = ({ props }) => {
-  return (
-    <section>
-      <div color="black">
-        <h1>Islands Section of Islands Page</h1>
-      </div>
-    </section>
-  );
-};
-
-export default Islands;
