--- src/Pages/marketplace/components/Cards.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-const Cards = ({ props }) => {
-  return (
-    <section>
-      <div color="black">
-        <h1>Card Section of Marketplace Page</h1>
-      </div>
-    </section>
-  );
-};
-
-export default Cards;
