--- src/Pages/propertyDetails/App.js
@@ -1,16 +1,0 @@
-import React, { useState } from "react";
-import PropertyDetails from "./components/PropertyDetails";
-import Navbar from "../../generalComponents/navbar/Navbar";
-import Footer from "../../generalComponents/footer/Footer";
-
-function PropertyDetailsApp() {
-  return (
-    <>
-      <Navbar />
-      <PropertyDetails />
-      <Footer />
-    </>
-  );
-}
-
-export default PropertyDetailsApp;
