--- src/Pages/propertyDetails/components/PropertyDetails.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-const PropertDetails = ({ props }) => {
-  return (
-    <section>
-      <div color="black">
-        <h1>Property Details Section of Propery details Page</h1>
-      </div>
-    </section>
-  );
-};
-
-export default PropertDetails;
