--- src/generalComponents/footer/Footer.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Footer = ({props}) => {
-  return <div>Footer component</div>;
-};
-
-export default Footer;
