--- src/generalComponents/navbar/Navbar.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Navbar = ({ props }) => {
-  return <nav>Navbar component</nav>;
-};
-
-export default Navbar;
